creating button add new cleaners
