+++ OCC/OCC/Views/Admin/DisplayCleanerList.cshtml
@@ -1,7 +1,7 @@
 ﻿@model IEnumerable<Cleaner>
-
 @*<link href="~/css/selectCleaner.css" rel="stylesheet" />*@
 <div class="our-webcoderskull">
+    <h3 class="card-title bg-info">Cleaners</h3>
     <table class="table table-striped table-bordered table-sm">
         <tr>
             <th class="text-right">CleanerId</th>
@@ -15,55 +15,101 @@
             <th>Morning</th>
             <th>Afternoon</th>
             <th>Night</th>
-            <th>Afternoon</th>
             <th>Weekends</th>
             <th>Actions</th>
         </tr>
 
 </div>
 @foreach (var item in Model)
-{<tr>
-    <td class="text-right">@item.CleanerId</td>
-    <td>@item.FirstName</td>
-    <td>@item.Email</td>
-    <td>@item.Location</td>
-    <td>@item.IsCleaner</td>
-    <td>@item.ExperienceLevel</td>
-    <td>@item.Certificate</td>
-    <td>@item.BankAccount</td>
-    <td>@item.Morning</td>
-    <td>@item.Afternoon</td>
-    <td>@item.Night</td>
-    <td>@item.Weekends</td>
-    <td class="text-center">
-        <form asp-action="Delete" method="post">
-            <input type="hidden" name="ProductId" value="@item.CleanerId" />
+{
+    @if (@item.IsCleaner == true)
+    {<tr>
+            <td class="text-right">@item.CleanerId</td>
+            <td>@item.FirstName</td>
+            <td>@item.Email</td>
+            <td>@item.Location</td>
+            <td>@item.IsCleaner</td>
+            <td>@item.ExperienceLevel</td>
+            <td>@item.Certificate</td>
+            <td>@item.BankAccount</td>
+            <td>@item.Morning</td>
+            <td>@item.Afternoon</td>
+            <td>@item.Night</td>
+            <td>@item.Weekends</td>
 
-            <a asp-action="Edit" class="btn btn-sm btn-warning" asp-route-productId="@item.CleanerId">
-                Edit
-            </a>
-            <button type="submit" class="btn btn-danger btn-sm">
-                Delete
-            </button>
-        </form>
-    </td>
-</tr>
+            <td class="text-center">
+                <form asp-action="Delete" method="post">
+                    <input type="hidden" name="ProductId" value="@item.CleanerId" />
+
+                    <a asp-action="Edit" class="btn btn-sm btn-warning" asp-route-productId="@item.CleanerId">
+                        Edit
+                    </a>
+                    <button type="submit" class="btn btn-danger btn-sm">
+                        Delete
+                    </button>
+                </form>
+            </td>
+        </tr>
+    }
 }
 </table>
+<div class="our-webcoderskull">
+    <h3 class="card-title bg-info">Potential Cleaners</h3>
+    <table class="table table-striped table-bordered table-sm">
+        <tr>
+            <th class="text-right">CleanerId</th>
+            <th>Name</th>
+            <th>Email</th>
+            <th>Location</th>
+            <th>IsCleaner</th>
+            <th>ExperienceLevel</th>
+            <th>Certificate</th>
+            <th>BankAccount</th>
+            <th>Morning</th>
+            <th>Afternoon</th>
+            <th>Night</th>
+            <th>Weekends</th>
+            <th>Actions</th>
+        </tr>
 
+</div>
 
-@*<tr>
-    <td class="text-right">@item.CleanerId</td>
-    <td>@item.FirstName</td>
-    <td>@item.Email</td>
-    <td>@item.Location</td>
-    <td>@item.IsCleaner</td>
-    <td>@item.ExperienceLevel</td>
-    <td>@item.Certificate</td>
-    <td>@item.BankAccount</td>
-    <td>@item.Morning</td>
-    <td>@item.Afternoon</td>
-    <td>@item.Night</td>
-    <td>@item.Weekends</td>
+@foreach (var item in Model)
+{
+    @if (@item.IsCleaner == false)
+    {
+        <tr>
+            <td class="text-right">@item.CleanerId</td>
+            <td>@item.FirstName</td>
+            <td>@item.Email</td>
+            <td>@item.Location</td>
+            <td>@item.IsCleaner</td>
+            <td>@item.ExperienceLevel</td>
+            <td>@item.Certificate</td>
+            <td>@item.BankAccount</td>
+            <td>@item.Morning</td>
+            <td>@item.Afternoon</td>
+            <td>@item.Night</td>
+            <td>@item.Weekends</td>
+            <td class="text-center">
+                <form asp-action="Delete" method="post">
+                    <input type="hidden" name="ProductId" value="@item.CleanerId" />
 
-</tr>*@
+                    <a asp-action="Edit" class="btn btn-sm btn-warning" asp-route-productId="@item.CleanerId">
+                        Edit
+                    </a>
+                    <button type="submit" class="btn btn-danger btn-sm">
+                        Delete
+                    </button>
+                </form>
+            </td>
+        </tr>
+    }
+}
+</table>
+<div class="d-flex justify-content-center">
+    <button type="submit" class="btn btn-success btn-lg">
+        Add New Cleaners
+    </button>
+</div>
+
