+++ OCC/OCC/Views/Admin/EditCleaners.cshtml
@@ -1,0 +1,34 @@
+﻿@model Cleaner
+
+<form asp-action="Edit" method="post">
+    <input type="hidden" asp-for="CleanerId" />
+
+    <div class="form-group">
+        <label asp-for="FirstName"></label>
+        <div><span asp-validation-for="FirstName" class="text-danger"></span></div>
+        <input asp-for="FirstName" class="form-control" />
+    </div>
+
+    <div class="form-group">
+        <label asp-for="Email"></label>
+        <div><span asp-validation-for="Email" class="text-danger"></span></div>
+        <textarea asp-for="Email" class="form-control"></textarea>
+    </div>
+
+    <div class="form-group">
+        <label asp-for="Location"></label>
+        <div><span asp-validation-for="Location" class="text-danger"></span></div>
+        <input asp-for="Location" class="form-control" />
+    </div>
+
+    <div class="form-group">
+        <label asp-for="IsCleaner"></label>
+        <div><span asp-validation-for="IsCleaner" class="text-danger"></span></div>
+        <input asp-for="IsCleaner" class="form-control" />
+    </div>
+
+    <div class="text-center">
+        <button class="btn btn-primary" type="submit">Save</button>
+        <a asp-action="DisplayCleanerList" class="btn btn-secondary">Cancel</a>
+    </div>
+</form>
